Added Dutch and Portuguese language options
