+++ src/routes/quiz/[language]/+page.js
@@ -1,5 +1,6 @@
-export async function load({ fetch }) {
-    const response = await fetch('/quiz');
+export async function load({ fetch, params }) {
+    const { language } = params; 
+    const response = await fetch(`/quiz/${language}`);
     const questions = await response.json();
 
     if (response.ok) {
