+++ src/routes/quiz/[language]/questions/Dutch.json
@@ -1,0 +1,12 @@
+[
+    {
+        "question": "Fiets",
+        "options": [ "Car", "Bus", "Bicycle", "Boat" ],
+        "answer": "Bicycle"
+    },
+    {
+        "question": "Tussendoortje",
+        "options": [ "Snack", "Interview", "Exam", "Traffic Jam" ],
+        "answer": "Snack"
+    }
+]
