+++ src/routes/quiz/[language]/questions/German.json
@@ -1,0 +1,52 @@
+[
+    {
+        "question": "Apfel",
+        "options": [ "Orange", "Apple", "Pear", "Banana" ],
+        "answer": "Apple"
+    },
+    {
+        "question": "Tante",
+        "options": [ "Uncle", "Aunt", "Niece", "Nephew" ],
+        "answer": "Aunt"
+    },
+    {
+        "question": "Straße",
+        "options": [ "Road", "River", "Building", "Park" ],
+        "answer": "Road"
+    },
+    {
+        "question": "Baum",
+        "options": [ "Car", "Tree", "Cat", "Sun" ],
+        "answer": "Tree"
+    },
+    {
+        "question": "Buch",
+        "options": [ "Pen", "Book", "Window", "Door" ],
+        "answer": "Book"
+    },
+    {
+        "question": "Fenster",
+        "options": [ "Wall", "Window", "Flower", "Sky" ],
+        "answer": "Window"
+    },
+    {
+        "question": "Tisch",
+        "options": [ "Chair", "Table", "Lamp", "Bed" ],
+        "answer": "Table"
+    },
+    {
+        "question": "Blume",
+        "options": [ "Rose", "Tulip", "Flower", "Daisy" ],
+        "answer": "Flower"
+    },
+    {
+        "question": "Katze",
+        "options": [ "Bird", "Cow", "Cat", "Horse" ],
+        "answer": "Cat"
+    },
+    {
+        "question": "Sonne",
+        "options": [ "Moon", "Star", "Sun", "Planet" ],
+        "answer": "Sun"
+    }
+]
